completed v1 functionality, adding/updating avatar
