+++ client/src/components/Nav.tsx
@@ -1,16 +1,45 @@
-import { Link } from "react-router-dom";
-import woofMathLogo from "../assets/woofmath_logo_1.png";
 import { CurrentApp, UserInfo } from "../types/types";
 import MenuIcon from "@mui/icons-material/Menu";
+import * as React from "react";
+import Menu from "@mui/material/Menu";
+import MenuItem from "@mui/material/MenuItem";
+import { useNavigate } from "react-router-dom";
 
 type NavTypes = {
   isLoggedIn: boolean;
+  setIsLoggedIn: React.Dispatch<React.SetStateAction<boolean>>;
   userInfo: UserInfo;
   currentApp: CurrentApp;
 };
 
 // TODO: STATE VALUE TO SET WHICH GAME THE USER IS ON + TO SET NAV TITLE (MATH OR READING)
-function Nav({ isLoggedIn, userInfo, currentApp }: NavTypes) {
+function Nav({ isLoggedIn, setIsLoggedIn, userInfo, currentApp }: NavTypes) {
+  const navigate = useNavigate();
+  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
+  const open = Boolean(anchorEl);
+  const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
+    setAnchorEl(event.currentTarget);
+  };
+  const handleClose = () => {
+    setAnchorEl(null);
+  };
+
+  const handleLogout = () => {
+    localStorage.removeItem("token");
+    localStorage.removeItem("userId");
+    // window.location.reload();
+    setIsLoggedIn(false);
+    navigate("/welcome");
+  };
+
+  const navHome = () => {
+    if (currentApp === "Woof Reading") {
+      navigate("/reading");
+    } else if (currentApp === "Woof Math") {
+      navigate("/math");
+    } else navigate("/");
+  };
+
   return (
     <>
       <nav
@@ -25,12 +54,37 @@ function Nav({ isLoggedIn, userInfo, currentApp }: NavTypes) {
           <img src={`../../avatars/${userInfo.avatar_name}.png`} alt="" />
         </div>
 
-        <div className="navTitle">{currentApp}</div>
+        <div className="navTitle" onClick={navHome}>
+          {currentApp}
+        </div>
 
         {isLoggedIn && userInfo && userInfo.username && (
-          <Link to={"/me"} className="navUser">
-            <MenuIcon />
-          </Link>
+          // <Link to={"/me"} className="navUser">
+          <div className="navMenu">
+            <MenuIcon
+              id="basic-button"
+              aria-controls={open ? "basic-menu" : undefined}
+              aria-haspopup="true"
+              aria-expanded={open ? "true" : undefined}
+              onClick={handleClick}
+            />
+            <Menu
+              id="basic-menu"
+              anchorEl={anchorEl}
+              open={open}
+              onClose={handleClose}
+              MenuListProps={{
+                "aria-labelledby": "basic-button",
+              }}
+            >
+              <MenuItem onClick={() => navigate("/me")}>My account</MenuItem>
+              <MenuItem onClick={() => navigate("/progress")}>
+                Progress dashboard
+              </MenuItem>
+              <MenuItem onClick={() => navigate("/")}>Change games</MenuItem>
+              <MenuItem onClick={handleLogout}>Logout</MenuItem>
+            </Menu>
+          </div>
         )}
       </nav>
     </>
