+++ client/src/index.css
@@ -756,7 +756,7 @@ button {
   display: flex;
   justify-content: start;
   align-items: center;
-  padding-left: 15px;
+  padding-left: 12px;
 }
 
 .navLogo > img {
@@ -773,7 +773,7 @@ button {
   /* font-family: "Ribeye"; */
 }
 
-.navUser {
+/* .navUser {
   display: flex;
   justify-content: end;
   align-items: center;
@@ -781,13 +781,18 @@ button {
   text-align: right;
 }
 
-.navUser img {
-  width: 55px;
-}
-
 .navUser > h4 {
   font-size: 17px;
+} */
+
+.navMenu {
+  display: flex;
+  justify-content: end;
+  align-items: center;
+  padding-right: 12px;
+  color: white;
 }
+
 /* ************************************************************ */
 /* SCOREBAR STYLING */
 
@@ -1821,6 +1826,23 @@ p {
   border: 1px solid rgb(223, 223, 223);
 }
 
+.accountAvatarContainer {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 95%;
+  /* border: 1px solid rgb(223, 223, 223); */
+  border-radius: 14px;
+  padding: 20px;
+  gap: 5px;
+  text-align: center;
+}
+
+.accountUserAvatar {
+  width: 200px;
+}
+
 .accountContentContainer > h3 {
   font-variant: small-caps;
   margin-top: 5px;
@@ -2679,6 +2701,10 @@ ul li {
   }
 
   .accountContentContainer {
+    width: 85%;
+  }
+
+  .accountAvatarContainer {
     width: 85%;
   }
